refactor: move align_contours_with_transform function to utils.py for better organization
